requirements.txt
discord.py
PyNaCl
asyncio